tweaks to clarify and add hint
Add explanations to 1.3 and 1.4 to clarify language. Some students didn't realize that adding code so the dog can walk to right meant making a new script that is triggered when right key is pressed.

Also added hint to 2.3 to look in sensing palette for the mouse x and mouse y blocks, as students haven't explored this category yet.